chore: delete duplicated files, move the test classes to the write folder.
