+++ src/test/java/aicore/unit/model/ModelChatFunctionalityTests.java
@@ -1,4 +1,4 @@
-package aicore.app.model;
+package aicore.unit.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
